VideoPress: Simplify preview-on-hover player customization
Derive the preview/playback customize payloads from a single helper so
the mirrored payloads cannot drift, and share the hover payload keys the
same way. Move the logic out of the view.ts webpack entry into a
side-effect-free preview-on-hover.ts module, which removes the
export-for-test seam and the @wordpress/dom-ready mock from the test.
Trim iframe API stubs the code under test never calls.

No behavior change; test assertions are unchanged.
